Fix the stop button
Added a Watcher class that is created in the CreateConnection context manager. This represents a second process that spawns when the context manager is entered. This process waits for the parent process to be killed without cleaning up. When the parent is killed it connects to the Create and shuts it off and disconnects.

Closes #9